add OrdersFilterBar refactor, UI improvements at Orders.
