+++ client/src/components/pages/Store/Orders/OrdersFilterBar/index.js
@@ -1,0 +1,1 @@
+export * from "./OrdersFilterBar";
